Reorganize component files
